Temp changes for DXC support. And 64 bit.

+++ Source/Launcher/Launcher_Win32.cpp
@@ -338,7 +338,7 @@ int APIENTRY WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
             DispatchMessage(&curMsg);
         }
     }
-    retVal = curMsg.wParam;
+    retVal = (int)curMsg.wParam;
 
     // Tear down windows
     application->DetachWindow(appWindow);
